AOSP security dialog
As pointed out by CopperheadOS on Twiiter[1], Android security updates
include the Android operating system updates in addition to closed
source binaries[2]. Since we cannot control whether or not these
binaries get updated (since it relies on both the vendors to provide
the patches and the OEM to distribute them), we should make it clear
that we are only merging what comes from Google through the Android
Open Source Project repositories. While some may consider this change
pedantic, it does not make it untrue or unreasonable to do.

This changes the look and feel of the 'Security level' preference to
only show it as "AOSP security patch level" but also launch a dialog
explaining the difference between AOSP security updates and actual
security updates. Furthermore, explain to the user that we merge in
functional updates along with the security updates.

[1]: https://twitter.com/CopperheadOS/status/984509294941892608
[2]: https://source.android.com/security/bulletin/2018-04-01#qualcomm-closed-source-components

Change-Id: Ia78de284277b39a176cb2c2d4146cf0f32f39644
Signed-off-by: Nathan Chancellor <nathan@dirtyunicorns.com>